requirements.txt
python-dotenv==1.0.1
streamlit==1.41.1
langchain==0.3.17
google-generativeai==0.8.4
langchain-google-genai==2.0.9
langchain-core==0.3.49

